Service_System/requirements.txt
django
djangorestframework
pillow
requests
